pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "leaflet>=0.0.3",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "qrcode>=8.1",
    "reportlab>=4.3.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
